Cleanup: remove dead code, extract create_websocket_client()
