+++ cli/src/perspectives.rs
@@ -1,7 +1,7 @@
 use graphql_client::{GraphQLQuery};
 use graphql_ws_client::graphql::StreamingOperation;
-use serde_json::{json, Value};
-use crate::{util::query, startup::get_executor_url};
+use serde_json::Value;
+use crate::util::{create_websocket_client, query};
 use anyhow::{Result, Context};
 use chrono::naive::NaiveDateTime;
 
@@ -9,11 +9,6 @@ type DateTime = NaiveDateTime;
 
 use self::all::AllPerspectives;
 use self::add_link::AddLinkPerspectiveAddLink;
-//use websocket::{ClientBuilder, Message};
-//use websocket::header::{Headers, Authorization};
-
-//use graphql_ws_client::{GraphQLClientClientBuilder};
-//use async_tungstenite::tungstenite::{client::IntoClientRequest, http::HeaderValue};
 
 #[derive(GraphQLQuery)]
 #[graphql(
@@ -152,66 +147,7 @@ pub async fn run_infer(cap_token: String, uuid: String, prolog_query: String) ->
         _ => v,
     })
 }
-/*
-mod schema {
-    cynic::use_schema!("../core/lib/src/schema.gql");
-}
 
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema_path = "../core/lib/src/schema.gql", graphql_type = "Link")]
-struct Link {
-    predicate: Option<String>,
-    source: String,
-    target: String,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema_path = "../core/lib/src/schema.gql", graphql_type = "ExpressionProof")]
-struct ExpressionProof {
-    invalid: Option<bool>,
-    key: String,
-    signature: String,
-    valid: Option<bool>,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema_path = "../core/lib/src/schema.gql", graphql_type = "LinkExpression")]
-struct LinkExpression {
-    author: String,
-    data: Link,
-    proof: ExpressionProof,
-    timestamp: String,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(schema_path = "../core/lib/src/schema.gql", graphql_type = "AgentStatus")]
-struct AgentStatus {
-    did: Option<String>,
-    did_document: Option<String>,
-    error: Option<String>,
-    is_initialized: bool,
-    is_unlocked: bool,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(
-    schema_path = "../core/lib/src/schema.gql",
-    graphql_type = "Subscription"
-)]
-struct AgentStatusChangedSubscription {
-    agent_status_changed: AgentStatus,
-}
-
-#[derive(cynic::QueryFragment, Debug)]
-#[cynic(
-    schema_path = "../core/lib/src/schema.gql",
-    graphql_type = "SubscriptionRoot"
-)]
-struct LinkAddedSubscription {
-    #[arguments(uuid = "uuid")]
-    perspective_link_added: Option<LinkExpression>,
-}
- */
 struct TokioSpawner(tokio::runtime::Handle);
 
 impl TokioSpawner {
@@ -234,10 +170,6 @@ impl futures::task::Spawn for TokioSpawner {
     }
 }
 
-//fn build_query() -> cynic::StreamingOperation<'static, LinkAddedSubscription> {
-//    use cynic::SubscriptionBuilder;
-//    LinkAddedSubscription::build(())
-//}
 
 #[derive(GraphQLQuery)]
 #[graphql(
@@ -248,145 +180,18 @@ impl futures::task::Spawn for TokioSpawner {
 pub struct SubscriptionLinkAdded;
 
 pub async fn run_watch(cap_token: String, id: String) -> Result<()> {
-    
-    use async_tungstenite::tungstenite::{client::IntoClientRequest, http::HeaderValue};
     use futures::StreamExt;
-    use graphql_ws_client::GraphQLClientClientBuilder;
 
-    let url = get_executor_url()?.replace("http", "ws");
-    let mut request = url.into_client_request().unwrap();
-    request.headers_mut().insert(
-        "Sec-WebSocket-Protocol",
-        HeaderValue::from_str("graphql-transport-ws").unwrap(),
-    );
-    request.headers_mut().insert(
-        "Authorization",
-        HeaderValue::from_str(&cap_token).unwrap(),
-    );
-    let (connection, _) = async_tungstenite::tokio::connect_async(request)
-        .await?;
-
-    println!("Connected");
-
-    let (sink, stream) = connection.split();
-
-
-    let mut client = GraphQLClientClientBuilder::new()
-        .build(stream, sink, TokioSpawner::current())
-        .await
-        .unwrap();
-
-    let op = StreamingOperation::<SubscriptionLinkAdded>::new(subscription_link_added::Variables {
+    let mut client = create_websocket_client(cap_token).await?;
+    let mut stream = client.streaming_operation(StreamingOperation::<SubscriptionLinkAdded>::new(subscription_link_added::Variables {
         uuid: id,
-    });
-
-    //let query_op = StreamingOperation::<QueryLinks>::new(query_links::Variables { 
-    //    uuid: id,
-    //    query: query_links::LinkQuery {
-    //        source: None, 
-    //        target: None, 
-    //        predicate: None,
-    //        fromDate: None,
-    //        untilDate: None,
-    //        limit: None,
-    //    }
-    //});
-
-    let mut stream = client.streaming_operation(op).await.unwrap();
+    })).await?;
     println!("Running subscription apparently?");
     while let Some(item) = stream.next().await {
         println!("{:?}", item);
     }
 
     println!("after loop");
-  
 
- /*
-    let (connection, _) = async_tungstenite::tokio::connect_async(request)
-        .await
-        .unwrap();
-
-    println!("Connected");
-
-    let (sink, stream) = connection.split();
-
-    let mut client = GraphQLClientClientBuilder::new()
-        .build(stream, sink, TokioSpawner::current())
-        .await
-        .unwrap();
-
-    let mut stream = client.streaming_operation(build_query()).await.unwrap();
-    println!("Running subscription apparently?");
-    while let Some(item) = stream.next().await {
-        println!("{:?}", item);
-    }
-  */
-    /*
-    let mut headers = Headers::new();
-    headers.set(Authorization(cap_token));
-    let executor_url = get_executor_url()?;
-    let mut client = ClientBuilder::new("ws://localhost:12000/graphql")
-        .unwrap()
-        .custom_headers(&headers) 
-        .add_protocol("graphql-transport-ws") 
-        .connect(None)
-        .unwrap();
-
-    client.send_message(&Message::text(serde_json::to_string(&json!({
-        "type": "connection_init"
-    }))?
-    )).unwrap();
-    let init_response = client.recv_message();
-    println!("Init response: {:?}", init_response);
-
-    //while let Ok(message) = client.recv_message() {
-    //    println!("After init: Received message: {:?}", message);
-    //}
-
-    /* 
-    let message = r#"
-    {
-        "id":"dc04f846-fb97-42b0-bfb8-7ddf0540aedf",
-        "type":"subscribe",
-        "payload":{
-            "variables":{
-                "uuid":"da0333e9-275d-4b57-8851-0d1678d75a1c",
-                "query":{
-                    "source":"ad4m://self",
-                    "predicate":"flux://has_channel"
-                }
-            },
-            "extensions":{},
-            "operationName": "perspectiveQueryLinks",
-            "query":"query perspectiveQueryLinks($uuid: String!, $query: LinkQuery!) {\n  perspectiveQueryLinks(query: $query, uuid: $uuid) {\n    author\n    timestamp\n    data {\n      source\n      predicate\n      target\n    }\n    proof {\n      valid\n      invalid\n      signature\n      key\n    }\n  }\n}\n"
-        }
-    }"#;
-    client.send_message(&Message::text(message)).unwrap();
-     
-    */
-    
-    //let gql_query = format!("perspectiveLinkAdded() {{ author timestamp data {{ source, predicate, target }} proof {{ valid, invalid, signature, key }} }}", id);
-    client.send_message(&Message::text(serde_json::to_string(&json!({
-        "type": "subscribe",
-        "id": "1",
-        "payload": {
-            "variables":{
-                "uuid": id,
-                //"query":{
-                //    "source":"ad4m://self"
-                //}
-            },
-            "extensions":{},
-            "operationName": "perspectiveLinkAdded",
-            "query": "query perspectiveLinkAdded($uuid: String!) { author timestamp data { source, predicate, target } proof { valid, invalid, signature, key } }"
-        },
-    }))?
-    )).unwrap(); 
-
-    while let Ok(message) = client.recv_message() {
-        println!("Received message: {:?}", message);
-    }
-        
- */
     Ok(())
 }
